+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `spending_keys` to wallet impl [#218] 
+
 ## [0.20.0] - 2023-11-01
 
 ### Changed
@@ -445,6 +449,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Implementation of `Store` trait from `wallet-core`
 - Implementation of `State` and `Prover` traits from `wallet-core`
 
+[#218]: https://github.com/dusk-network/wallet-cli/issues/218
 [#214]: https://github.com/dusk-network/wallet-cli/issues/214
 [#210]: https://github.com/dusk-network/wallet-cli/issues/210
 [#98]: https://github.com/dusk-network/wallet-cli/issues/98